pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "llmtestgen"
version = "0.1.0"
description = "Automated LLM test generator with local web interface"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name="RICHELET Arthur", email="richelar@mail.uc.com" },
  { name="DERACHE Cedric", email="derachcc@mail.uc.com"}
]
dependencies = [
"fastapi==0.115.0",
"uvicorn[standard]==0.30.1",
"pydantic==2.7.1",
"python-dotenv==1.0.1",
"httpx==0.27.0",
"typer==0.20.0",
"gitpython==3.1.45",
"markdown-it-py==3.0.0"
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "mypy",
]

[project.scripts]
llmtestgen = "src.cli:main"
llmtestgen-setup = "src.cli:setup"
llmtestgen-settings = "src.cli:settings"

[tool.pytest.ini_options]
markers = [
    "live_llm: marks tests that hit real LLM APIs (requires --live-llm)",
    "live_git: marks tests that hit real Git repositories (requires --live-git)",
]

[tool.bandit]
skips = ["B101"]
